Move LSP command to experimental as bundle-lsp
The command is now at `databricks experimental bundle-lsp` instead of
`databricks bundle lsp`. This better reflects its experimental status.
